% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_TPMs.R
\name{extract_salmon_quants}
\alias{extract_salmon_quants}
\title{Extract TPMS and number of reads from salmon generated quant.sf files}
\usage{
extract_salmon_quants(
  organism1,
  organism2,
  analysis_dir,
  outdir,
  regexp = STAR_SALMON_REGEXP
)
}
\arguments{
\item{organism1}{the first organism of the coral}

\item{organism2}{the second organism of the coral}

\item{analysis_dir}{the analysis directory}

\item{outdir}{the output directory}

\item{regexp}{the regular expression to match files. Default
value: "*/*salmon_quant/quant.sf"}
}
\description{
\code{extract_salmon_quants()} extracts all the salmon_quant.sf files
in the given analysis directory and writes 2 output files, 1 for
the TPMs and one for the number of reads
}
